chages the directory for time interval resolver, adds EOF
