+++ app/assets/stylesheets/shared/_distribution-summary.scss
