+++ app/views/development_metrics/merge_time/_main_metric.erb
