+++ app/views/pull_requests/index.erb
@@ -32,7 +32,7 @@
   </div>
   <div class="row">
     <div class="col-md-10 offset-1 mt-5">
-      <%@intervals_pull_requests.each do |k,v|%>
+      <%@pull_requests.each do |k,v|%>
         <h6><%= k %> hours:</h6>
         <ul>
           <% v.each do |pr_link| %>
